fix: pos of the new tab for opts 'do not move it' and 'use general rule'
For 'use general rule', original openerTabId was used instead of using
the general rule.

For 'do not move it', original openerTabId was used which moved
the new tab horizontally (indenting it), creating the branch. Now,
the flat list of tabs will be produced, which will emulate Firefox
positioning for child tabs.